completed 80% of using promise flow

+++ src/js/directives/chips.js
@@ -1,5 +1,5 @@
 (function() {
-    angular.module('angular.chips')
+    angular.module('angular.chips', [])
         .directive('chips', Chips)
         .controller('chipsController', ChipsController);
 
@@ -25,11 +25,37 @@
         }
     }
 
+    function deferChip(data, promise) {
+        var defer = {};
+
+        defer.data = data;
+        defer.isLoading = false;
+        defer.isFailed = false;
+
+        if (promise) {
+            defer.isLoading = true;
+            promise.then(function(data) {
+                defer.data = data;
+                defer.isLoading = false;
+            }, function() {
+                defer.isLoading = false;
+                defer.isFailed = true;
+            });
+        }
+
+        return defer;
+    }
+
     function Chips($compile, $timeout, DomUtil) {
+
         function linkFun(scope, iElement, iAttrs, ngModelCtrl, transcludefn) {
             if ((error = validation(iElement)) !== undefined) {
                 throw error;
             }
+
+            var model = ngModel(ngModelCtrl);
+            var isDeferFlow = iAttrs.hasOwnProperty('defer');
+
             /*
              *  @scope.chips.addChip should be called by chipControl directive or custom XXXcontrol directive developed by end user
              *  @scope.chips.deleteChip will be called by removeChip directive
@@ -41,13 +67,21 @@
              */
             scope.chips.list;
 
-            var model = ngModel(ngModelCtrl);
-
 
             scope.chips.addChip = function(data) {
-                var updatedData, loadingObjRef;
+                var updatedData;
                 scope.render !== undefined ? updatedData = scope.render({ data: data }) : updatedData = data;
-                isPromiseLike(updatedData) ? updatedData.then(update) : update(updatedData);
+                // isPromiseLike(updatedData) ? deferChip(updatedData).update(data) : update(updatedData);
+
+                if (isPromiseLike(updatedData)) {
+                    updatedData.then(function(response) {
+                        model.add(response);
+                    });
+                    scope.chips.list.push(deferChip(data, updatedData));
+                    scope.$apply();
+                } else {
+                    update(updatedData);
+                }
 
                 function update(data) {
                     scope.chips.list.push(data);
@@ -56,7 +90,9 @@
             };
 
             scope.chips.deleteChip = function(index) {
-                scope.chips.list.splice(index, 1);
+                var deletedChip = scope.chips.list.splice(index, 1);
+                if (typeof deletedChip !== 'string' && deletedChip.isFailed)
+                    return;
                 model.delete(index);
             }
 
@@ -64,7 +100,17 @@
              * ngModel values are copied when it's updated outside
              */
             ngModelCtrl.$render = function() {
-                scope.chips.list = angular.copy(ngModelCtrl.$modelValue);
+                if (isDeferFlow) {
+                    var index, list = [];
+                    for (index = 0; index < ngModelCtrl.$modelValue.length; index++) {
+                        // list.push(ngModelCtrl.$modelValue[index]);
+                        list.push(deferChip(ngModelCtrl.$modelValue[index]))
+                    }
+                    scope.chips.list = list;
+                } else {
+                    scope.chips.list = angular.copy(ngModelCtrl.$modelValue);
+                }
+
             }
 
             var chipNavigate = null;
@@ -79,10 +125,17 @@
                     return index;
                 }
             }
-            /*Below code will extract the chip-tmpl and compile inside the chips directive scope*/
+
+            /*Extract the chip-tmpl and compile inside the chips directive scope*/
             var rootDiv = angular.element('<div></div>');
             var tmpl = iElement.find('chip-tmpl').remove();
+            var chipTextNode, chipBindedData, chipBindedDataSuffix;
+
+            if (isDeferFlow)
+                DomUtil(tmpl).attachDataObjToTextNode();
+
             tmpl.attr('ng-repeat', 'chip in chips.list track by $index');
+            tmpl.attr('ng-class', '{\'chip-failed\':chip.isFailed}')
             tmpl.attr('tabindex', '-1')
             tmpl.attr('index', '{{$index+1}}')
             rootDiv.append(tmpl);
